Feat: added login/logout and view products with no CUD functions

+++ index.php
@@ -1,5 +1,10 @@
 <?php
+session_start();
 include 'database.php';
+include 'check_auth.php';
+include 'database.php';
+
+$is_authenticated = isAuthenticated();
 
 // Pagination settings
 $records_per_page = 10;
@@ -224,8 +229,10 @@ $result = $conn->query($sql);
                             <th scope="col" style="width: 80px;">ID</th>
                             <th scope="col">Product Name</th>
                             <th scope="col" style="width: 150px;">Price</th>
-                            <th scope="col" style="width: 120px;">Quantity</th>
-                            <th scope="col" style="width: 200px;" class="text-center">Actions</th>
+                            <?php if ($is_authenticated): ?>
+                                <th scope="col" style="width: 120px;">Quantity</th>
+                                <th scope="col" style="width: 200px;" class="text-center">Actions</th>
+                            <?php endif; ?>
                         </tr>
                     </thead>
                     <tbody>
@@ -235,36 +242,40 @@ $result = $conn->query($sql);
                                     <td><?= htmlspecialchars($row['id']) ?></td>
                                     <td class="product-name"><?= htmlspecialchars($row['product_name']) ?></td>
                                     <td>₱<?= number_format($row['product_price'], 2) ?></td>
-                                    <td><?= htmlspecialchars($row['quantity']) ?></td>
-                                    <td class="text-center">
-                                        <div class="btn-group-actions">
-                                            <a href="update.php?id=<?= $row['id'] ?>" class="btn btn-primary btn-sm">
-                                                <svg xmlns="http://www.w3.org/2000/svg" width="14" height="14" fill="currentColor" class="bi bi-pencil" viewBox="0 0 16 16">
-                                                    <path d="M12.146.146a.5.5 0 0 1 .708 0l3 3a.5.5 0 0 1 0 .708l-10 10a.5.5 0 0 1-.168.11l-5 2a.5.5 0 0 1-.65-.65l2-5a.5.5 0 0 1 .11-.168zM11.207 2.5 13.5 4.793 14.793 3.5 12.5 1.207zm1.586 3L10.5 3.207 4 9.707V10h.5a.5.5 0 0 1 .5.5v.5h.5a.5.5 0 0 1 .5.5v.5h.293zm-9.761 5.175-.106.106-1.528 3.821 3.821-1.528.106-.106A.5.5 0 0 1 5 12.5V12h-.5a.5.5 0 0 1-.5-.5V11h-.5a.5.5 0 0 1-.468-.325" />
-                                                </svg>
-                                                <span class="btn-text">Edit</span>
-                                            </a>
-                                            <a href="delete.php?id=<?= $row['id'] ?>" class="btn btn-danger btn-sm"
-                                                onclick="return confirm('Are you sure you want to delete this product?');">
-                                                <svg xmlns="http://www.w3.org/2000/svg" width="14" height="14" fill="currentColor" class="bi bi-trash" viewBox="0 0 16 16">
-                                                    <path d="M5.5 5.5A.5.5 0 0 1 6 6v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5m2.5 0a.5.5 0 0 1 .5.5v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5m3 .5a.5.5 0 0 0-1 0v6a.5.5 0 0 0 1 0z" />
-                                                    <path d="M14.5 3a1 1 0 0 1-1 1H13v9a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V4h-.5a1 1 0 0 1-1-1V2a1 1 0 0 1 1-1H6a1 1 0 0 1 1-1h2a1 1 0 0 1 1 1h3.5a1 1 0 0 1 1 1zM4.118 4 4 4.059V13a1 1 0 0 0 1 1h6a1 1 0 0 0 1-1V4.059L11.882 4zM2.5 3h11V2h-11z" />
-                                                </svg>
-                                                <span class="btn-text">Delete</span>
-                                            </a>
-                                        </div>
-                                    </td>
+                                    <?php if ($is_authenticated): ?>
+                                        <td><?= htmlspecialchars($row['quantity']) ?></td>
+                                        <td class="text-center">
+                                            <div class="btn-group-actions">
+                                                <a href="update.php?id=<?= $row['id'] ?>" class="btn btn-primary btn-sm">
+                                                    <svg xmlns="http://www.w3.org/2000/svg" width="14" height="14" fill="currentColor" class="bi bi-pencil" viewBox="0 0 16 16">
+                                                        <path d="M12.146.146a.5.5 0 0 1 .708 0l3 3a.5.5 0 0 1 0 .708l-10 10a.5.5 0 0 1-.168.11l-5 2a.5.5 0 0 1-.65-.65l2-5a.5.5 0 0 1 .11-.168zM11.207 2.5 13.5 4.793 14.793 3.5 12.5 1.207zm1.586 3L10.5 3.207 4 9.707V10h.5a.5.5 0 0 1 .5.5v.5h.5a.5.5 0 0 1 .5.5v.5h.293zm-9.761 5.175-.106.106-1.528 3.821 3.821-1.528.106-.106A.5.5 0 0 1 5 12.5V12h-.5a.5.5 0 0 1-.5-.5V11h-.5a.5.5 0 0 1-.468-.325" />
+                                                    </svg>
+                                                    <span class="btn-text">Edit</span>
+                                                </a>
+                                                <a href="delete.php?id=<?= $row['id'] ?>" class="btn btn-danger btn-sm"
+                                                    onclick="return confirm('Are you sure you want to delete this product?');">
+                                                    <svg xmlns="http://www.w3.org/2000/svg" width="14" height="14" fill="currentColor" class="bi bi-trash" viewBox="0 0 16 16">
+                                                        <path d="M5.5 5.5A.5.5 0 0 1 6 6v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5m2.5 0a.5.5 0 0 1 .5.5v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5m3 .5a.5.5 0 0 0-1 0v6a.5.5 0 0 0 1 0z" />
+                                                        <path d="M14.5 3a1 1 0 0 1-1 1H13v9a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V4h-.5a1 1 0 0 1-1-1V2a1 1 0 0 1 1-1H6a1 1 0 0 1 1-1h2a1 1 0 0 1 1 1h3.5a1 1 0 0 1 1 1zM4.118 4 4 4.059V13a1 1 0 0 0 1 1h6a1 1 0 0 0 1-1V4.059L11.882 4zM2.5 3h11V2h-11z" />
+                                                    </svg>
+                                                    <span class="btn-text">Delete</span>
+                                                </a>
+                                            </div>
+                                        </td>
+                                    <?php endif; ?>
                                 </tr>
                             <?php endwhile; ?>
                         <?php else: ?>
                             <tr>
-                                <td colspan="5" class="text-center empty-state">
+                                <td colspan="<?= $is_authenticated ? '5' : '3' ?>" class="text-center empty-state">
                                     <div class="py-5">
                                         <svg xmlns="http://www.w3.org/2000/svg" width="48" height="48" fill="currentColor" class="bi bi-inbox text-muted mb-3" viewBox="0 0 16 16">
                                             <path d="M4.98 4a.5.5 0 0 0-.39.188L1.54 8H6a.5.5 0 0 1 .5.5 1.5 1.5 0 1 0 3 0A.5.5 0 0 1 10 8h4.46l-3.05-3.812A.5.5 0 0 0 11.02 4zm-1.17-.437A1.5 1.5 0 0 1 4.98 3h6.04a1.5 1.5 0 0 1 1.17.563l3.7 4.625a.5.5 0 0 1 .106.374l-.39 3.124A1.5 1.5 0 0 1 14.117 13H1.883a1.5 1.5 0 0 1-1.489-1.314l-.39-3.124a.5.5 0 0 1 .106-.374z" />
                                         </svg>
                                         <p class="text-muted mb-3">No products found</p>
-                                        <a href="create.php" class="btn btn-theme">Add Your First Product</a>
+                                        <?php if ($is_authenticated): ?>
+                                            <a href="create.php" class="btn btn-theme">Add Your First Product</a>
+                                        <?php endif; ?>
                                     </div>
                                 </td>
                             </tr>
